Several changes: - Add systemc testbenches adapted from gemmini spike tests - Add script gen_testbenches.py to generate aforementions testbenches - Add sim_infra headers copied from gemmini spike tests which are needed by the systemc testbenches - Add simulator CMakeLists.txt to overwrite the default generated one - Add test_runner.sh utils script - Add gemmini_testbench.h simulator testbench header - Add setup_sim_folder.py script - Add call to setup_sim_folder.py in main.cc

+++ sim_infra/rocc-software/src/xcustom.h
@@ -152,8 +152,8 @@
                                                           \
     opcode      = OPCODE_CUSTOM3;                                  \
     funct_write = func7;                                           \
-    rs1         = (uint64_t) rs1_val;                                          \
-    rs2         = rs2_val;                                          \
+    rs1         = (uint64_t) (rs1_val);                                          \
+    rs2         = (uint64_t) (rs2_val);                                          \
     wait(1, SC_SEC);                                            \
   } \
